Removing the last bit of dependency injection
