+++ test/index.js
@@ -22,16 +22,12 @@ import is from 'is';
 import through2 from 'through2';
 
 import {Firestore} from '../src/index';
-import {referencePkg} from '../src/reference';
+import {DocumentReference, CollectionReference} from '../src/reference';
 import {ResourcePath} from '../src/path';
 import {createInstance} from '../test/util/helpers';
 
 const gax = require('google-gax');
 const grpc = new gax.GrpcClient().grpc;
-
-const reference = referencePkg(Firestore);
-const DocumentReference = reference.DocumentReference;
-const CollectionReference = reference.CollectionReference;
 
 const PROJECT_ID = 'test-project';
 const DATABASE_ROOT = `projects/${PROJECT_ID}/databases/(default)`;
